requirements_gui.txt
python-docx>=0.8.11
tkinter
threading
json
os
time
datetime
shutil
